Refactor FileFullPath
